requirements.txt
streamlit
pillow
tensorflow
opencv-python
numpy